perf(ci): Optimize CI pipeline and resolve code quality issues
Optimize CI/CD pipeline with comprehensive caching and improve code quality
by resolving all clang-tidy warnings.

CI/CD Optimizations:
- Add LLVM/Clang 18 installation caching to reduce setup time
- Add Homebrew package caching for macOS builds
- Improve Docker build cache scoping for better reusability
- Optimize macOS workflow to run after Linux tests complete
- Remove redundant format checks from macOS workflow

Code Quality Improvements:
- Fix clang-tidy warnings for unused variables and parameters
- Eliminate magic numbers by introducing named constants (kBytesPerMB)
- Resolve variable shadowing in RAII guard classes
- Fix dead store warnings with NOLINT annotations where appropriate
- Add move constructors/operators delete for RAII guards
- Improve code formatting consistency

Bug Fixes:
- Initialize variables to prevent undefined behavior
- Add missing #include <memory> in mygram-cli.cpp
- Make ExtractAllFilters static as it doesn't use instance state
- Fix potential overflow in result sorting with proper casting

+++ src/query/result_sorter.cpp
@@ -192,9 +192,7 @@ std::vector<DocId> ResultSorter::SortAndPaginate(std::vector<DocId>& results, co
   // Check for overflow in offset + limit and clamp to results.size()
   uint64_t total_needed_64 = static_cast<uint64_t>(query.offset) + static_cast<uint64_t>(query.limit);
   // Clamp to min(total_needed_64, results.size()) to avoid out-of-bounds access
-  size_t total_needed = (total_needed_64 > results.size())
-                            ? results.size()
-                            : static_cast<size_t>(total_needed_64);
+  size_t total_needed = (total_needed_64 > results.size()) ? results.size() : static_cast<size_t>(total_needed_64);
 
   // Use partial_sort aggressively when total_needed is significantly smaller than result size
   // Threshold: if we need less than 50% of results, use partial_sort
@@ -206,7 +204,8 @@ std::vector<DocId> ResultSorter::SortAndPaginate(std::vector<DocId>& results, co
     // For 800K results with K=100: ~800K * log2(100) ≈ 5.3M operations
     // vs full sort: 800K * log2(800K) ≈ 15.9M operations
     // Memory: in-place, no additional allocation
-    std::partial_sort(results.begin(), results.begin() + total_needed, results.end(), comparator);
+    std::partial_sort(results.begin(), results.begin() + static_cast<std::ptrdiff_t>(total_needed), results.end(),
+                      comparator);
 
     spdlog::trace("Used partial_sort for {} out of {} results", total_needed, results.size());
   } else {
